Move ObjectHelper namespace
