+++ src/PSRule/Pipeline/PipelineReciever.cs
@@ -1,5 +1,5 @@
 ﻿using Newtonsoft.Json;
-using PSRule.Commands;
+using PSRule.Runtime;
 using System.Collections;
 using System.Collections.Generic;
 using System.IO;
